Type changes for Java

+++ io/github/kitswas/VGP_Data_Exchange/GameButtons.java
@@ -4,25 +4,29 @@ package io.github.kitswas.VGP_Data_Exchange;
 // https://learn.microsoft.com/en-us/uwp/api/windows.gaming.input.gamepadbuttons
 // Adapted from windows.gaming.input.h
 public enum GameButtons {
-  None(0x0),
-  Menu(0x1),
-  View(0x2),
-  A(0x4),
-  B(0x8),
-  X(0x10),
-  Y(0x20),
-  DPadUp(0x40),
-  DPadDown(0x80),
-  DPadLeft(0x100),
-  DPadRight(0x200),
-  LeftShoulder(0x400),
-  RightShoulder(0x800),
-  LeftThumbstick(0x1000),
-  RightThumbstick(0x2000);
+    None((short) 0x0),
+    Menu((short) 0x1),
+    View((short) 0x2),
+    A((short) 0x4),
+    B((short) 0x8),
+    X((short) 0x10),
+    Y((short) 0x20),
+    DPadUp((short) 0x40),
+    DPadDown((short) 0x80),
+    DPadLeft((short) 0x100),
+    DPadRight((short) 0x200),
+    LeftShoulder((short) 0x400),
+    RightShoulder((short) 0x800),
+    LeftThumbstick((short) 0x1000),
+    RightThumbstick((short) 0x2000);
 
-  private final int value;
+    private final short value;
 
-  GameButtons(int value) { this.value = value; }
+    GameButtons(short value) {
+        this.value = value;
+    }
 
-  public int getValue() { return value; }
+    public short getValue() {
+        return value;
+    }
 }
